+++ ray.h
@@ -5,12 +5,12 @@
 #include "constants.h"
 
 struct Ray {
-	glm::dvec3 O, D;
+	glm::vec3 O, D;
 
-	Ray(glm::dvec3 O_in, glm::dvec3 D_in):
+	Ray(glm::vec3 O_in, glm::vec3 D_in):
 		O(O_in), D(D_in) {}
 };
 
-glm::mat3 orthonormalZ(const glm::dvec3& z);
+glm::mat3 orthonormalZ(const glm::vec3& z);
 
-glm::dvec3 normalFromUV(const glm::dvec3& u, const glm::dvec3& v);
+glm::vec3 normalFromUV(const glm::vec3& u, const glm::vec3& v);
